feat: Add draw line feature
Adds the feature to draw a line on the canvas.
All of the line's points must be within the canvas
for it to be drawn. A line can be drawn using the
`draw` command with proceeding x y co-ordinates e.g.
`draw 2 3` or `draw 2 3 F` where `F` would be the
desired pixel.