feat: Add area feature
Adds the feature to draw an area on the canvas.
All of the area's points must be within the canvas
for it to be drawn. An area can be drawn using the
`area` command with proceeding x1 y1 x2 y2 co-ordinates e.g.
`area 2 3 5 5` or `area 2 3 5 5 A` where `A` would be the
desired pixel.