feat: Add fill feature
Adds the feature to fill an area on the canvas.
The starting point of the fill must be within the
canvas for it to be filled.

An area can be drawn using the `fill` command with
proceeding x y co-ordinates e.g.
`fill 2 3` or `fill 2 3 F` where `F` would be the
desired pixel.

Please note the `fill` only fills pixel above/below
or to the sides and does not fill diagonal squares.